DjangoAndGeminiIntregration/requirements.txt
google-generativeai
python-dotenv
djangorestframework
djangorestframework-simplejwt
django-cors-headers